build: Support non-linux compilation
