+++ Makefile
@@ -4,6 +4,7 @@ GO ?= go
 GCC ?= gcc
 DOCKER ?= docker
 IMAGE_TAG ?= paulinhu/go-apparmor:1
+BUILD_TAGS ?= apparmor
 
 CWD := $(realpath .)
 OUTDIR := $(CWD)/build
@@ -19,6 +20,10 @@ image:
 
 .PHONY: build
 build:
+	$(GO) build -tags $(BUILD_TAGS) ./...
+
+.PHONY: example
+example:
 	pushd example/code && \
 	$(GO) build -ldflags '$(LDFLAGS)' -o $(OUTDIR)/$(BINARY) ./main.go || \
 	popd
