+++ pkg/apparmor/apparmor.go
@@ -1,8 +1,6 @@
 package apparmor
 
 import (
-	"runtime"
-
 	"github.com/go-logr/logr"
 )
 
@@ -26,19 +24,4 @@ type aa interface {
 
 	// LoadPolicy loads an AppArmor policy into the kernel.
 	LoadPolicy(fileName string) error
-}
-
-var goOS = func() string {
-	return runtime.GOOS
-}
-
-// NewAppArmor creates a new instance of the apparmor API.
-func NewAppArmor() aa {
-	if goOS() == "linux" {
-		return &AppArmor{
-			logger: logr.Discard(),
-		}
-	}
-
-	return &unsupported{}
 }
