+++ pkg/apparmor/apparmor_test.go
@@ -1,3 +1,6 @@
+//go:build linux && apparmor
+// +build linux,apparmor
+
 package apparmor
 
 import (
