+++ pkg/apparmor/apparmor_unsupported.go
@@ -1,0 +1,8 @@
+//go:build !linux || !apparmor
+// +build !linux !apparmor
+
+package apparmor
+
+func NewAppArmor() aa {
+	return &unsupported{}
+}
